requirements.txt
functions_framework
asttokens==2.4.1
Brotli==1.0.9
cachetools==4.2.4
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.0.4
colorama==0.4.6
cryptography==38.0.2
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
google-api-core==1.14.1
google-auth==1.21.3
google-cloud-bigquery==1.22.0
google-cloud-core==1.1.0
google-crc32c==1.1.2
google-resumable-media==0.5.1
googleapis-common-protos==1.63.0
idna==3.6
importlib-metadata==4.11.4
Shapely
numpy
packaging==23.2
pip==24.0
platformdirs==4.0.0
proto-plus==1.23.0
protobuf==3.20.*
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
pytz==2024.1
pyu2f==0.1.5
requests==2.31.0
rsa==4.9
setuptools==69.0.3
six==1.16.0
stack-data==0.6.2
tqdm==4.66.2
typing_extensions==4.7.1
urllib3==2.2.1
wheel==0.42.0
win-inet-pton==1.1.0
zipp==3.15.0
scikit-learn
